Fix Shutdown
- Split Code into separate files
- Fix issue where script will not stop on Kodi exit, crashing Kodi and/or delaying shutdown by 20+ secs.
